+++ tools/onnx-graphsurgeon/tests/onnx_models.py
@@ -2,6 +2,7 @@ from onnx_graphsurgeon.logger.logger import G_LOGGER
 from onnx_graphsurgeon.ir.tensor import Tensor, ConstantTensor, VariableTensor
 from onnx_graphsurgeon.ir.graph import Graph
 from onnx_graphsurgeon.ir.node import Node
+from onnx_graphsurgeon.importers.onnx_importer import OnnxImporter
 
 G_LOGGER.severity = G_LOGGER.ULTRA_VERBOSE
 
@@ -15,11 +16,12 @@ import os
 TEST_ROOT = os.path.realpath(os.path.dirname(__file__))
 
 class Model(object):
-    def __init__(self, path: str, inputs: List[Tensor], outputs: List[Tensor], nodes: List[Node]):
+    def __init__(self, path: str, inputs: List[Tensor], outputs: List[Tensor], nodes: List[Node], opset: int):
         self.path = path
         self.inputs = inputs
         self.outputs = outputs
         self.nodes = nodes
+        self.opset = opset
 
     def load(self):
         return onnx.load(self.path)
@@ -49,16 +51,31 @@ class Model(object):
 
 
 def identity_model():
+    path = os.path.join(TEST_ROOT, "models", "identity.onnx")
+    model = onnx.load(path)
+
     x = VariableTensor(name="x", dtype=np.float32, shape=(1, 1, 2, 2))
     y = VariableTensor(name="y", dtype=np.float32, shape=(1, 1, 2, 2))
     node = Node(op="Identity", inputs=[x], outputs=[y])
 
-    return Model(os.path.join(TEST_ROOT, "models", "identity.onnx"), inputs=[x], outputs=[y], nodes=[node])
+    return Model(path, inputs=[x], outputs=[y], nodes=[node], opset=OnnxImporter.get_opset(model))
+
+
+def dim_param_model():
+    path = os.path.join(TEST_ROOT, "models", "dim_param.onnx")
+    model = onnx.load(path)
+
+    x = VariableTensor(name="Input:0", dtype=np.float32, shape=("dim0", 16, 128))
+    y = VariableTensor(name="Output:0", dtype=np.float32, shape=("dim0", 16, 128))
+    node = Node(op="Identity", inputs=[x], outputs=[y])
+
+    return Model(path, inputs=[x], outputs=[y], nodes=[node], opset=OnnxImporter.get_opset(model))
 
 
 def lstm_model():
     path = os.path.join(TEST_ROOT, "models", "lstm.onnx")
-    onnx_graph = onnx.load(path).graph
+    model = onnx.load(path)
+    onnx_graph = model.graph
 
     def load_initializer(index: int) -> np.ndarray:
         return onnx.numpy_helper.to_array(onnx_graph.initializer[index])
@@ -79,12 +96,12 @@ def lstm_model():
     attrs["hidden_size"] = 5
     node = Node(op="LSTM", attrs=attrs, inputs=[X, W, R, B, Tensor.empty(), Tensor.empty(), initial_c], outputs=[Y, Y_h, Y_c])
 
-    return Model(path, inputs=[X, W, R, B, initial_c], outputs=[Y, Y_h, Y_c], nodes=[node])
+    return Model(path, inputs=[X, W, R, B, initial_c], outputs=[Y, Y_h, Y_c], nodes=[node], opset=OnnxImporter.get_opset(model))
 
 
 def scan_model():
     path = os.path.join(TEST_ROOT, "models", "scan.onnx")
-    onnx_graph = onnx.load(path).graph
+    model = onnx.load(path)
 
     # Body graph
     sum_in = VariableTensor(name="sum_in", dtype=np.float32, shape=(2, ))
@@ -112,4 +129,4 @@ def scan_model():
     attrs["body"] = body_graph
     attrs["num_scan_inputs"] = 1
     scan_node = Node(op="Scan", inputs=inputs, outputs=outputs, attrs=attrs)
-    return Model(path, inputs=inputs, outputs=outputs, nodes=[scan_node])
+    return Model(path, inputs=inputs, outputs=outputs, nodes=[scan_node], opset=OnnxImporter.get_opset(model))